refactors the UserForm component, and fixes test file to accomidate changes
