+++ src/styles/UserForm.less
@@ -1,3 +1,5 @@
+@import './variables.less';
+
 .userForm {
   .userButton {
     display: flex;
